Move song and album data to JSON
