+++ src/api/albums.json
@@ -1,0 +1,44 @@
+[
+  {
+    "id": 0,
+    "name": "Top 50 Global",
+    "image": "img8",
+    "desc": "Your weekly update of the most played tracks",
+    "bgColor": "#2a4365"
+  },
+  {
+    "id": 1,
+    "name": "Top 50 India",
+    "image": "img9",
+    "desc": "Your weekly update of the most played tracks",
+    "bgColor": "#22543d"
+  },
+  {
+    "id": 2,
+    "name": "Trending India",
+    "image": "img10",
+    "desc": "Your weekly update of the most played tracks",
+    "bgColor": "#742a2a"
+  },
+  {
+    "id": 3,
+    "name": "Trending Global",
+    "image": "img16",
+    "desc": "Your weekly update of the most played tracks",
+    "bgColor": "#44337a"
+  },
+  {
+    "id": 4,
+    "name": "Mega Hits,",
+    "image": "img11",
+    "desc": "Your weekly update of the most played tracks",
+    "bgColor": "#234e52"
+  },
+  {
+    "id": 5,
+    "name": "Happy Favorites",
+    "image": "img15",
+    "desc": "Your weekly update of the most played tracks",
+    "bgColor": "#744210"
+  }
+]
