+++ src/api/songs.json
@@ -1,0 +1,66 @@
+[
+  {
+    "id": 0,
+    "name": "Song One",
+    "image": "img1",
+    "file": "song1",
+    "desc": "Put a smile on your face with these happy tunes",
+    "duration": "3:00"
+  },
+  {
+    "id": 1,
+    "name": "Song Two",
+    "image": "img2",
+    "file": "song2",
+    "desc": "Put a smile on your face with these happy tunes",
+    "duration": "2:20"
+  },
+  {
+    "id": 2,
+    "name": "Song Three",
+    "image": "img3",
+    "file": "song3",
+    "desc": "Put a smile on your face with these happy tunes",
+    "duration": "2:32"
+  },
+  {
+    "id": 3,
+    "name": "Song Four",
+    "image": "img4",
+    "file": "song1",
+    "desc": "Put a smile on your face with these happy tunes",
+    "duration": "2:50"
+  },
+  {
+    "id": 4,
+    "name": "Song Five",
+    "image": "img5",
+    "file": "song2",
+    "desc": "Put a smile on your face with these happy tunes",
+    "duration": "3:10"
+  },
+  {
+    "id": 5,
+    "name": "Song Six",
+    "image": "img14",
+    "file": "song3",
+    "desc": "Put a smile on your face with these happy tunes",
+    "duration": "2:45"
+  },
+  {
+    "id": 6,
+    "name": "Song Seven",
+    "image": "img7",
+    "file": "song1",
+    "desc": "Put a smile on your face with these happy tunes",
+    "duration": "2:18"
+  },
+  {
+    "id": 7,
+    "name": "Song Eight",
+    "image": "img12",
+    "file": "song2",
+    "desc": "Put a smile on your face with these happy tunes",
+    "duration": "2:35"
+  }
+]
