+++ src/components/DisplayAlbum/DisplayAlbum.jsx
@@ -1,5 +1,7 @@
 import React, { useContext } from 'react'
-import { albumsData, assets, songsData } from '../../assets/assets'
+import albumsData from '../../api/albums.json'
+import songsData from '../../api/songs.json'
+import { assets } from '../../assets/assets'
 import { useParams } from 'react-router-dom'
 import { PlayerContext } from '../../context/PlayerContext';
 import DisplayNav from '../DisplayNav/DisplayNav';
@@ -14,7 +16,7 @@ const DisplayAlbum = () => {
         <>
             <DisplayNav />
             <div className='mt-10 flex gap-8 flex-col md:flex-row md:items-end'>
-                <img className='w-48 rounded' src={albumData.image} alt="" />
+                <img className='w-48 rounded' src={assets[albumData.image]} alt="" />
                 <div className='flex flex-col'>
                     <p>Playlist</p>
                     <h2 className='text-5xl font-bold mb-4 md:text-7xl'>{albumData.name}</h2>
@@ -34,7 +36,7 @@ const DisplayAlbum = () => {
                     <div onClick={() => playWithId(item.id)} className='grid grid-cols-3 sm:grid-cols-4 gap-2 p-2 items-center text-[#a7a7a7] hover:bg-[#ffffff2b] cursor-pointer' key={index}>
                         <p className='text-white'>
                             <b className='mr-4 text-[#a7a7a7]'>{index + 1}</b>
-                            <img className='inline w-10 mr-5' src={item.image} alt="" />
+                            <img className='inline w-10 mr-5' src={assets[item.image]} alt="" />
                             {item.name}
                         </p>
                         <p className='text-[15px]'>{albumData.name}</p>
